Facebook Login now Works - but unpleasantly
When launched app will ask to build IIS server, let it do so.  If the server is properly set up on your computer you will be able to go to the logon page and log in to Facebook.  If you are already logged into Facebook in the browser you are using, you will instantly get redirected to your FBProfile page - a page that displays your Facebook name and Facebook ID.  This is thoroughly tested to work when everything is properly set up, if you have problems with it then something is wrong with your server settings and don't worry about the Facebook stuff for the time.  In fact, the Facebook stuff currently interferes with being able to use other site functionality (profiles, properly logging in to check out, etc).  It is just a proof of concept for now. (all the profile stuff still works in the background, you just have to not have a logged in Facebook account cookie active in your browser - site currently does not have a "Log Off" button.)

Signed-off-by: ecnalyr <ecnalyr@gmail.com>