Changed CSS to test ability to do so.
Pages look fine but if you log in the top bar becomes three lines thick instead of one and this makes every other page unreadable because it covers the top 'two lines' worth of each page.  Also, when you are logged in, the text that says "Welcome User!" is the same color as the black background of the top bar, so you cannot see it unless you highlight the text.

Signed-off-by: ecnalyr <ecnalyr@gmail.com>