Added randomly generated 7 digit key to be used as redemption code.
When a user confirms their purchase a 7 digit code is generated.  This will be the code a cashier would input to see if the customer actually purchased a drink.  This key currently displays on the user's checkout-confirmation screen.  To get see an example of the key, click on the POS link on the navigation bar.

Also added the bool "Active" to the Order table so that we can inactivate codes when they are redeemed.

Signed-off-by: ecnalyr <ecnalyr@gmail.com>